\name{RunKODAMAmatrix}
\alias{RunKODAMAmatrix}
\alias{S3method(RunKODAMAmatrix, default)}
\alias{S3method(RunKODAMAmatrix, Seurat)}
\alias{S3method(RunKODAMAmatrix, giotto)}
\alias{S3method(RunKODAMAmatrix, SingleCellExperiment)}
\alias{S3method(RunKODAMAmatrix, SpatialExperiment)}
\title{Documentation for RunKODAMAmatrix}
\description{See source file for full details.}
